# src/shared/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API User",
    "description": "Api responsible for creating and validating users",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "Develop",
      "email": "augustodanieltec@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    },
    {
      "url": "http://portifolio-user-api.herokuapp.com",
      "description": "Heroku server"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/users/getUsers": {
      "get": {
        "summary": "Get all users",
        "description": "Get all exist users in database",
        "operationId": "getUsers",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Users was retrieved successfully!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No token provided"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Users not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/getUser/{email}": {
      "get": {
        "summary": "Get user by email",
        "description": "This route is responsible for get user by email",
        "operationId": "getUserByEmail",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "User email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {},
          "400": {},
          "401": {},
          "404": {},
          "500": {}
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}